refactor: 패키지명 어댑터로 변경
