+++ src/main/java/roomescape/adapter/AdminController.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.adapter;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,5 +14,10 @@ public class AdminController {
     @GetMapping("/admin/reservation")
     public String manageReservation() {
         return "admin/reservation-legacy";
+    }
+
+    @GetMapping("/admin/time")
+    public String manageReservationTime() {
+        return "admin/time";
     }
 }
