+++ src/main/java/roomescape/adapter/ReservationTime/ReservationTimeController.java
@@ -1,4 +1,4 @@
-package roomescape.interface_adapter.ReservationTime;
+package roomescape.adapter.ReservationTime;
 
 import java.util.List;
 import org.springframework.http.ResponseEntity;
