+++ src/main/java/roomescape/adapter/ReservationTime/ReservationTimeRepositoryImpl.java
@@ -1,4 +1,4 @@
-package roomescape.interface_adapter.ReservationTime;
+package roomescape.adapter.ReservationTime;
 
 import java.sql.PreparedStatement;
 import java.time.LocalTime;
